+++ classes/local/thumbnail_course_lock.php
@@ -1,0 +1,37 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+namespace local_course_banner_builder\local;
+
+defined('MOODLE_INTERNAL') || die();
+
+/**
+ * Concurrency boundary for a course-thumbnail lifecycle transition.
+ *
+ * @package    local_course_banner_builder
+ * @copyright  2026 Kevin Jarniac
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+interface thumbnail_course_lock {
+    /**
+     * Execute one callback while holding the lock for a course.
+     *
+     * @param int $courseid
+     * @param callable $callback
+     * @return mixed
+     */
+    public function execute(int $courseid, callable $callback): mixed;
+}